requirements.txt
gradio
requests
faiss-cpu
sentence-transformers
